compressor_lifetime_4/backend/requirements.txt
fastapi>=0.116.0
uvicorn>=0.30.0
websockets>=12.0
pydantic>=2.8.0
PyYAML>=6.0.2
nidaqmx>=1.0.2
pytest>=8.3.3
pytest-asyncio>=0.24.0
httpx>=0.28.1
